requirements.txt
streamlit
pandas
altair
pillow